pyproject.toml
[project]
name = "modelviewprojection"
version = "0.0.2"
authors = [{name="William Emerison Six", email="billsix@gmail.com"}]
description = "ModelViewProjection - Tutorial for learning 3D Programming"

dependencies = [
    "autoflake==2.3.1",
    "glfw==2.9.0",
    "imgui==2.0.0",
    "isort==6.0.1",
    "numpy==2.2.6",
    "PyOpenGL==3.1.9",
    "ruff==0.11.11",
    "pytest==8.3.5",
    "spyder==6.0.7",
    "staticlocal==0.0.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest]
testpaths = ["tests"]